Fix: Arrays can be printed and modify_field standalone (#5848)
### Description
A previous update forced the `T` under `Object<T>` to implement
`UpcastObject<dyn Any>` to be able to print it.
However, this trait was never implemented on objects. Arrays can be
converted to objects but it's not possible to cast them down to arrays
again, because arrays are traits in Rust (dynamic pointers).
Until we wrap arrays with a newtype to cast them back to a regular
struct, we can fix the Dafny runtime so that they are still printed as
"object"

Second, modify-field!() at the end of a block was triggering parsing
issues. I added a block wrapper in the macro to prevent that issue.

### How has this been tested?
Two tests added in the runtime.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>