fix: Make FuncExtensions internal again, handle it in DafnyGeneratedFromDafnyPost.py instead (#5860)
### Description
#5757 made `FuncExtensions` in the C# runtime public instead of
internal, which is not ideal since it increases the API surface area of
the runtime and is a one-way door (and causes false positives in
prerelease regressions testing). It was done because it seemed to be
necessary, but the actual root cause was dropping these extension
methods when post-processing Dafny's output. This PR reverts the change
and special-cases the class in the python script instead.

### How has this been tested?
Existing tests.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Mikael Mayer <mimayere@amazon.com>