fix: Update parameter resolution in new resolver (#5884)
The old and new resolvers had got out of sync in how they resolve
parameters, especially with regard to `nameonly` parameters and
parameters with default values. Apparently, a change in this behavior
had been rolled into the old resolver, but not the new. This PR brings
the new resolver up to date.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>