Chore: Help message for when nightly tests fail (#5869)
### Description
Whenever nightly tests fails, some tests are failing but it's not
obvious how to fix nightly tests. There is some documentation but I
think it's nicer for developers to know
1. If there is currently an action that is likely to resolve it soon.
2. If not, what action they can take.

### How has this been tested?
Investigating this during a nightly test failure. It gives the following
helpful clickable message:
> Error: Last run of nightly-build.yml did not succeed:
https://github.com/dafny-lang/dafny/actions/runs/11562412876
A run of nightly-build.yml is currently in_progress:
https://github.com/dafny-lang/dafny/actions/runs/11576162026, just
re-run this test once it is finished.


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>