fix: Omit type antecedents for unassigned variables (#5877)
In let expressions and named function results, the bound variable was
previously introduced with an assumption that the variable satisfied its
type. That is not sound, since the type may be empty. This PR removes
those premature type assumptions.

Fixes #1958

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>