Fix issue 5554 soundness (#5874)
### Description
This PR updates the Dafny CS runtime to use BigIntegers instead of int32
when retrieving the number of elements in a multiset. As this number may
be over INT_MAX, this change avoids potential overflows. This is similar
to what is done in the Java runtime.

### How has this been tested?
A test was added:
Source/IntegrationTests/TestFiles/LitTests/LitTest/git-issues/git-issue-5554.dfy

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>