docs: eliminate mention of "function method" from "Getting Started" guide
Dafny functions (since Dafny 4.0 or so) are by default what previously was called "function method".
The previous default can be achieved with ghost functions.
In PR #5006 it was decided to not distract learners with that distinction already in the "Getting Started" guide
and to only introduce them to non-ghost functions there.