docs: eliminate mention of "function method" from "Getting Started" guide (#5944)
Dafny functions (since Dafny 4.0 or so) are by default what previously
was called "function method". The previous default can be achieved with
ghost functions. In
https://github.com/dafny-lang/dafny/pull/5006#issuecomment-1973670652 it
was decided to not distract learners with that distinction already in
the "Getting Started" guide and to only introduce them to non-ghost
functions there.

This change eliminates a mention of "function method" that was probably
overlooked back then.

### Description
- Docs-only change. (Thus user-visible, but only in the documentation.)
- Inconsistency present also in the latest release, thus this change
should be backported to that.

- Tested by running `nix run nixpkgs#jekyll -- server --future` and then
navigating to `http://127.0.0.1:4000/docs/OnlineTutorial/guide` and
looking at the "Here we use `nats` […]" paragraph.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>