Fix ownership object issues (#5933)
Fixes #5932 

### Description
Previously, there was some conversion code that, in the case of an
upcast conversion, was removing a borrow annotation to ensure the object
stayed owned. This was done in the context of newtypes that had copy
semantics. However, that same piece of code is invoked in cases of
variables with non-copy semantics.
This PR fixes that by doing a proper check.

### How has this been tested?
The test traits.dfy has been extended, was not passing, and now passes.


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>