fix: Regenerate system module in DafnyRuntimeGo-gomod, add missing CI check (#5965)
### Description
When we split off the separate `DafnyRuntimeGo-gomod` copy of the Go
runtime, we neglected to duplicate the CI mechanism we have in each
runtime that ensures the generated `System` module matches what Dafny is
currently generating. As a result this copy was missing the
`CompanionStruct_Nat_.Is_` function Dafny now generates for subset types
like `nat`.

### How has this been tested?
Updated `runtime-tests.yml`.