Stop unnecessarily verifying symbols multiple times when using refinement (#5967)
The IDE requires defining symbols in the front-end that can be verified,
and matching these to verification tasks that are generated from the
back-end. The only reliable way of doing that matching is by comparing
original source locations. To prevent verifying the same task multiple
times, currently there is a `symbolsToVerify =
symbolsToVerify.DistinctBy(s => s.Tok)` call that prevents calling
verification for the same location multiple times. However, this is
unreliable in particular when working with module refinement.

### Description
- Let CliCompilation call `compilation.VerifyLocation` instead of
`compilation.VerifyCanVerify`, since the IDE does this as well and this
change forces the two to be more consistent.
- Prevent surfacing copied nodes as CanVerify targets, so 
  - No risk of verifying the same symbol twice
- There is no longer a need for doing `symbolsToVerify =
symbolsToVerify.DistinctBy(s => s.Tok)`, which could cause the wrong
CanVerify to be used.
- Let symbols redeclared in refining modules have their own source
location as origin, so they do not show up as "copies"

### How has this been tested?
- Update expect files

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>