fix: Concurrent.MutableMap doesn't work on complex key types in Go (#5984)
### What was changed?
Just wrapping a native concurrency-safe map implementation as we do for
the other languages doesn't work here, because Dafny cannot customize
the hashcode/equality functions used.

Addressed this by replacing the implementation with a mutex-guarded
`_dafny.Map` instead. This means most operations will be `O(n)` instead
of `O(1)`, just as they are on pure immutable Dafny `set` and `map`
values, but this can be improved in the future.

### How has this been tested?
Added a missing case to `ConcurrentExamples.dfy` that exposes this
issue, since Dafny strings are never translated directly to native
string values but a custom type instead.