Fixed RELEASE.md
The release process stated at the end to increment the patch version for the next release, but also that the release number "may have changed" at the beginning when it "should have changed". Moreover, the instructions on how to update the doo files and other goodies were missing. Added them from my notes.