Fix: Rust traits downcast now clone their arguments (#5993)
Fixes #5992

### What was changed?
When upcasting/downcasting, we often need a cloned reference. However,
until recently, the references were cloned only if the call site
required an owned referenced. Recently, references were borrowed and the
borrow converted to a clone only if the ownership of the reference was
necessary. However, this was implemented for Upcast, not downcast.
Hence, if the same reference needed multiple downcasts, the generated
Rust code would fail.

This PR fixes that by reusing the cloning code for upcasting but for
downcasting as well.

### How has this been tested?
Modified comp/rust/traits.dfy to account for two test cases that were
failing for a user.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>