Replace `INode.Tok` with `INode.Origin.Center` (#5983)
### What was changed?
- Replace `INode.Tok` with `INode.Origin.Center`. In many places this
means we can work with a single origin, instead of both `.Tok` and an
`.Origin`.
- Remove `BoogieRangeToken`
- "Hidden functions:" tooltip code was improved to reduce duplicates.
The new code prevents any overlapping ranges in reporting tooltips.
- Some untested behavior in
`ReporterExtensions.CreateDiagnosticRelatedInformationFor` was turned
off.

### How has this been tested?
Some tests have been updated for small improvements in behavior:
- "Argument to attribute Compile must be a literal" is reported on the
argument center, instead of on the attribute's `(`.
- For parallel assignment, `var a, b := <expr1>, <expr2>`, messages
about the whole assignment where shown on the `,`, while for regular
assignment these are shown on `:=`. Now they're always shown on the
`:=`.
- Diagnostics about a method now point to the method's name, instead of
the keyword `method`.
- When using `--show-snippets`, the diagnostics are now reported at the
same position as when not using `--show-snippets`.
- When an assertion fails to prove, the error range in the IDE includes
the `assert ` keyword, so it visually is different from the precondition
of the condition itself not proving.
- Locations and ranges reported by proof dependency logic code have been
made consistent with other diagnostics.

Some tests have different but basically the same behavior:
- "Assertion with {:only} temporarily transforms other assertions into
assumptions" is reported on the "o" of "only" instead of on the `{`
- `--json-diagnostics` now more often includes an "end" location
- Diagnostics about `(0 decreases to 1)` are now shown on the `d`
instead of the `(`

Some tests have different but still not desired behavior:
- "wrong number of arguments (got 2, but attribute 'Compile' expects at
most 1", is reported at the `@` instead of at the final `)`. This will
be improved by future item (4), so it will be reported on the `(`
instead of the `@`
- "incorrect argument type for attribute parameter '0' (expected bool,
found string) is now reported on the `@` instead of the final `)`. This
will be improved by future item (3), so it will be reported on the
center of the argument instead of the `@`


### Future
1. Inline usages of `TokenNode.Tok` and `RangeNode.Tok`
1. Merge TokenNode and RangeNode into Node by making TokenNode's
behavior (compute the range but only if needed) the one and only
1. Address comment `/* TODO should be b.Actual.Origin */`
1. Address comment `// TODO use e.Origin instead of e.Lhs.Origin`
1. No longer let Dafny.Token inherit from IOrigin
1. No longer let IOrigin inherit from Boogie.IToken


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>