Fix name clash in generated CS files when module and types have same name (#6019)
### What was changed?
This change updates the CS code generated from Dafny to fix name clashes
between namespace and class when a module defines a datatype with the
same name. On the example for
https://github.com/dafny-lang/dafny/issues/6014, we have:

```
module A {

  datatype A = Foo. // Notice the name of the datatype is the name of the module

}

module B {

   import opened A

   const bar: A

   method Main() {
     print "Hello!\n";
  }
}

```

which generated the code (simlified for readability)

```
namespace A {

  public interface _IA {
    bool is_Foo { get; }
    _IA DowncastClone();
  }
  public class A : _IA {
    public A() {
    }
    // [...]

    public static _IA create() {
      return new A();
    }

    private static readonly A._IA theDefault = create();

   // [...]

    }
  }
} // end of namespace A

namespace B {

  public partial class __default {
    public static void _Main(Dafny.ISequence<Dafny.ISequence<Dafny.Rune>> __noArgsParameter)
    {
      Dafny.Helpers.Print((Dafny.Sequence<Dafny.Rune>.UnicodeFromString("Hello!\n")).ToVerbatimString(false));
    }
    public static A._IA bar { get {
      return A.A.Default();
    } }
  }
} // end of namespace B

```


The expression `A._IA theDefault` does not compile in C#, as the first
A. is resolved to be the name of the class and not the name of the
namespace. One solution could be to ensure we have `_IA` here instead of
`A._IA` as for the `create` method, but the code to generate this type
name is used at other places where the `A._IA` is required (e.g in `
bar` method in `namespace B`).

This change changes the name of the namespace `A` to `_NA` when there is
a name clash with a datatype declared in `A`.

### How has this been tested?
Updated test gith-issues-6014.dfy


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Mikaël Mayer <MikaelMayer@users.noreply.github.com>